internal: Separate interface definitions from concrete implementations
